requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
peewee==3.17.0
pydantic
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
